Add pytest and split package
