+++ src/core/tools/lang.py
@@ -2,7 +2,7 @@ import javalang
 import re
 
 
-class MethodNode:
+class JavaAstNode:
     def __init__(self, name='', start_pos=0, end_pos=0):
         self.name = name
         self.start_pos = start_pos
@@ -14,7 +14,7 @@ class MethodNode:
         return self.name == '' and self.start_pos == 0 and self.end_pos == 0
 
     def __str__(self):
-        return "MethodNode(name={}, start_pos={}, end_pos={}, highlight=\n{})".format(self.name, self.start_pos, self.end_pos, self.highlight_line_numbers)
+        return "JavaAstNode(name={}, start_pos={}, end_pos={}, highlight=\n{})".format(self.name, self.start_pos, self.end_pos, self.highlight_line_numbers)
 
     def add_highlight_line_number(self, line_number):
         self.highlight_line_numbers.append(line_number)
@@ -38,13 +38,6 @@ def is_comment_line(line):
     return re.match(r'^(//|/\*|\*|\*/)', striped_line)
 
 
-def is_line_countable(line):
-    striped_line = line.strip()
-    is_comment = re.match(r'^(//|/\*|\*|\*/)', striped_line)
-    has_multi_chars = len(striped_line) > 1
-    return not is_comment and has_multi_chars
-
-
 def filter_ast_nodes_by_types(root, node_types):
     filtered_nodes = []
     for node in root:
@@ -60,7 +53,7 @@ def filter_ast_nodes_by_types(root, node_types):
     return filtered_nodes
 
 
-def load_file_methods(file_path):
+def filter_ast_nodes(file_path):
     method_nodes = []
     with open(file_path, 'r') as file:
         text = file.read()
@@ -69,7 +62,7 @@ def load_file_methods(file_path):
     filtered_nodes = filter_ast_nodes_by_types(
         tree, ['MethodDeclaration', 'ConstructorDeclaration', 'ClassDeclaration', 'EnumDeclaration', 'InterfaceDeclaration'])
     for filtered_node in filtered_nodes:
-        method_nodes.append(MethodNode(filtered_node.name,
+        method_nodes.append(JavaAstNode(filtered_node.name,
                             filtered_node.position.line))
 
     return method_nodes
@@ -104,7 +97,7 @@ def cal_method_end_pos(method_node, file_lines):
 
 
 def parse_method_metainfo(file_path):
-    method_nodes = load_file_methods(file_path)
+    method_nodes = filter_ast_nodes(file_path)
     file_lines = read_file_lines(file_path)
 
     for i in range(len(method_nodes)):
@@ -129,7 +122,7 @@ def load_patch_code_snippets(file_path, line_numbers):
     for m in method_nodes:
         print(m)
 
-    most_related_method = MethodNode()
+    most_related_method = JavaAstNode()
     for m in method_nodes:
         if len(m.highlight_line_numbers) > 0:
             if most_related_method.is_empty() or len(m.highlight_line_numbers) > len(most_related_method.highlight_line_numbers):
