+++ src/runner.py
@@ -1,0 +1,4 @@
+from core.large_language_models.codex import execute
+
+if __name__ == "__main__":
+    execute()
